+++ pkg/skr/nfs/gcpnfsvolume/loadPersistenceVolume.go
@@ -1,4 +1,4 @@
-package crgcpnfsvolume
+package gcpnfsvolume
 
 import (
 	"context"
