+++ pkg/skr/nfs/gcpnfsvolume/reconciler.go
@@ -1,0 +1,62 @@
+package gcpnfsvolume
+
+import (
+	"context"
+	cloudresourcesv1beta1 "github.com/kyma-project/cloud-manager/api/cloud-resources/v1beta1"
+	"github.com/kyma-project/cloud-manager/pkg/composed"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/klog/v2"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/cluster"
+)
+
+type Reconciler struct {
+	composedStateFactory composed.StateFactory
+	stateFactory         StateFactory
+}
+
+func (r *Reconciler) Run(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
+	state := r.newState(req.NamespacedName)
+	action := r.newAction()
+
+	return composed.Handle(action(ctx, state))
+}
+
+func (r *Reconciler) newState(name types.NamespacedName) *State {
+	return r.stateFactory.NewState(
+		r.composedStateFactory.NewState(name, &cloudresourcesv1beta1.GcpNfsVolume{}),
+	)
+}
+
+func (r *Reconciler) newAction() composed.Action {
+	return composed.ComposeActions(
+		"crGcpNfsVolumeMain",
+		composed.LoadObj,
+		composed.ComposeActions(
+			"crGcpNfsVolumeValidateSpec",
+			validateIpRange, validateCapacity, validateFileShareName),
+		addFinalizer,
+		loadKcpNfsInstance,
+		loadPersistenceVolume,
+		modifyKcpNfsInstance,
+		deletePersistenceVolume,
+		deleteKcpNfsInstance,
+		removeFinalizer,
+		removePersistenceVolumeFinalizer,
+		createPersistenceVolume,
+		modifyPersistenceVolume,
+		updateStatus,
+		composed.StopAndForgetAction,
+	)
+}
+
+func NewReconciler(kymaRef klog.ObjectRef, kcpCluster cluster.Cluster, skrCluster cluster.Cluster) Reconciler {
+	compSkrCluster := composed.NewStateCluster(skrCluster.GetClient(), skrCluster.GetEventRecorderFor("cloud-resources"), skrCluster.GetScheme())
+	compKcpCluster := composed.NewStateCluster(kcpCluster.GetClient(), kcpCluster.GetEventRecorderFor("cloud-control"), kcpCluster.GetScheme())
+	composedStateFactory := composed.NewStateFactory(compSkrCluster)
+	stateFactory := NewStateFactory(kymaRef, compKcpCluster, compSkrCluster)
+	return Reconciler{
+		composedStateFactory: composedStateFactory,
+		stateFactory:         stateFactory,
+	}
+}
